Fixed Delete Error and More Organized

+++ EmployeeAPI2/EmployeeAPI2/Controllers/EmpController.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-
+using System.Data.Entity;
 
 namespace EmployeeAPI2.Controllers
 {
@@ -16,7 +16,8 @@ namespace EmployeeAPI2.Controllers
         {
             return View();
         }
-
+        #region Add Employee
+        //Adds new Employees
         [Route("AddEmployee")]
         [HttpPost] 
         public async Task<ActionResult<List<NewEmp>>> AddEmp(NewEmp emp)
@@ -25,18 +26,25 @@ namespace EmployeeAPI2.Controllers
             await _context.SaveChangesAsync();
             return Ok(await _context.Employee.ToListAsync());
         }
+        #endregion
 
-        [Route("Delete")]   
+        #region Delete Employees
+        //Deletes Employees
+        [Route("DeleteEmployee")]   
         [HttpDelete]
 
-        public async Task<ActionResult> DeleteEmp(int id)
+        public async Task<ActionResult<List<NewEmp>>> DeleteEmp(int id)
         {
             var employee = await _context.Employee.FindAsync(id);
+            
+            if(employee == null)
+            {
+                return BadRequest("Employee Does Not Exist In Database");
+            }
             _context.Employee.Remove(employee);
             await _context.SaveChangesAsync();
             return Ok(await _context.Employee.ToListAsync());
         }
-        
-        
+        #endregion
     }
 }
